+++ superset-frontend/src/components/Select/Select.stories.tsx
@@ -17,127 +17,48 @@
  * under the License.
  */
 import React from 'react';
-import { useArgs } from '@storybook/client-api';
-import { OptionTypeBase } from 'react-select';
-import Select from '.';
-
-const OPTIONS = [
-  { label: 'Blue', value: 'blue' },
-  { label: 'Red', value: 'red' },
-  { label: 'Orange', value: 'orange' },
-];
+import Select, { SelectProps, OptionsPromise } from './Select';
 
 export default {
   title: 'Select',
-  argTypes: {
-    options: {
-      type: 'select',
-      options: OPTIONS,
-    },
-    multi: {
-      type: 'boolean',
-    },
-    value: {
-      type: 'string',
-    },
-    clearable: {
-      type: 'boolean',
-    },
-    placeholder: {
-      type: 'string',
-    },
-  },
+  component: Select,
 };
 
-export const SelectGallery = ({ value }: { value: OptionTypeBase }) => (
-  <>
-    <h4>With default value</h4>
-    <Select
-      value={OPTIONS[0]}
-      ignoreAccents={false}
-      name="select-datasource"
-      onChange={() => {}}
-      options={OPTIONS}
-      placeholder="choose one"
-      width={600}
-    />
-    <hr />
-    <h4>With no value</h4>
-    <Select
-      ignoreAccents={false}
-      name="select-datasource"
-      onChange={() => {}}
-      options={OPTIONS}
-      placeholder="choose one"
-      width={600}
-      value={value}
-    />
-    <hr />
-    <h4>Multi select</h4>
-    <Select
-      ignoreAccents={false}
-      name="select-datasource"
-      onChange={() => {}}
-      options={OPTIONS}
-      placeholder="choose one or more values"
-      width={600}
-      value={[OPTIONS[0]]}
-      multi
-    />
-  </>
-);
+export const InteractiveSelect = (args: SelectProps) => <Select {...args} />;
 
-SelectGallery.args = {
-  value: '',
-  options: OPTIONS,
-};
+async function fetchUserList(username: string): OptionsPromise {
+  console.log('fetching user', username);
 
-SelectGallery.story = {
-  parameters: {
-    knobs: {
-      disabled: true,
-    },
-  },
-};
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const InteractiveSelect = (args: any) => {
-  const [{ value, multi, clearable, placeholder }, updateArgs] = useArgs();
-  const onSelect = (selection: {}) => {
-    const { value }: { value?: any } = selection || {};
-    if (multi) {
-      updateArgs({ value: selection });
-      return;
-    }
-    updateArgs({ value });
-  };
-
-  return (
-    <Select
-      clearable={clearable}
-      onChange={onSelect}
-      name="interactive-select"
-      options={OPTIONS}
-      placeholder={placeholder}
-      with={600}
-      value={value}
-      multi={multi}
-    />
-  );
-};
+  return fetch('https://randomuser.me/api/?results=5')
+    .then(response => response.json())
+    .then(body =>
+      body.results.map(
+        (user: {
+          name: { first: string; last: string };
+          login: { username: string };
+        }) => ({
+          label: `${user.name.first} ${user.name.last}`,
+          value: user.login.username,
+        }),
+      ),
+    );
+}
 
 InteractiveSelect.args = {
-  value: '',
-  multi: false,
-  options: OPTIONS,
-  clearable: false,
-  placeholder: "I'm interactive",
+  showSearch: false,
+  options: fetchUserList,
+};
+
+InteractiveSelect.argTypes = {
+  mode: {
+    control: { type: 'select', options: ['single', 'multiple', 'tags'] },
+  },
 };
 
 InteractiveSelect.story = {
   parameters: {
     knobs: {
-      disabled: true,
+      disable: true,
     },
   },
 };
